requirements.txt
jinja2
jsonref
openapi_core
pydantic
sphinx
ruamel.yaml
jsonschema
bs4[spec]
markdownify[spec]
wsgi_intercept[spec]
oslotest[spec]
nova[spec]
glance[spec]
neutron[spec]
keystone[spec]
cinder[spec]
octavia[spec]
